You don't want a prolonged suspension, do you?
Grammar and spelling adjustments, make it clear to users that continuing to violate the same rule(s) is going to get them a prolonged suspension. Maybe this will help prevent *some* rulebreaking. Not going to add the lee-way for this because then they'll just think they're in the safe zone after 3 months.

Would like some feedback on this. It's already in the staff guidelines, I thought we should add it here so people don't ask why they got a 3 point or complain about how there's nothing in the rules that states we can give them a 3 point for a minor violation.