Prevent monetary offers
In an effort to protect users, removing the ability for there to be monetary offers. This is just using the idea that buying/selling stuff is not what the forum is for and any reports on it will not be handled.
Mimics the SYC shops rule, however disallows buying/selling